refactor: Price를 common 도메인으로 이동
