+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -12,7 +12,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-import kitchenpos.domain.Price;
+import kitchenpos.domain.common.Price;
 
 @Table(name = "menu")
 @Entity
