+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -7,7 +7,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import kitchenpos.domain.Price;
+import kitchenpos.domain.common.Price;
 
 @Table(name = "product")
 @Entity
