+++ src/test/java/kitchenpos/domain/menu/MenuValidatorTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.Price;
+import kitchenpos.domain.common.Price;
 import kitchenpos.domain.product.Product;
 import kitchenpos.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
