+++ src/test/java/kitchenpos/domain/menu/PriceTest.java
@@ -3,7 +3,7 @@ package kitchenpos.domain.menu;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Price;
+import kitchenpos.domain.common.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
